Optimised flow to submit API events to Pulsar (#982)
* Add optimized code for submitting several API events to Pulsar

* Add optimized code for submitting several API events to Pulsar

* Use optimized code to submit several API events

* Refactoring

* Remove unused code

* Send SubmitApiEvents Pulsar messages async.
