+++ internal/armada/server/submit_to_log.go
@@ -9,6 +9,7 @@ import (
 	"github.com/apache/pulsar-client-go/pulsar"
 	"github.com/gogo/protobuf/proto"
 	"github.com/gogo/protobuf/types"
+	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -27,9 +28,6 @@ import (
 	"github.com/G-Research/armada/pkg/armadaevents"
 	"github.com/G-Research/armada/pkg/client/queue"
 )
-
-// Id used for messages for which we can't use the kubernetesId.
-const LEGACY_RUN_ID = "00000000000000000000000000"
 
 // PulsarSubmitServer is a service that accepts API calls according to the original Armada submit API
 // and publishes messages to Pulsar based on those calls.
@@ -566,9 +564,70 @@ func (srv *PulsarSubmitServer) GetQueueInfo(ctx context.Context, req *api.QueueI
 	return srv.SubmitServer.GetQueueInfo(ctx, req)
 }
 
+// SubmitApiEvents converts several api.EventMessage into Pulsar state transition messages and publishes those to Pulsar.
+func (srv *PulsarSubmitServer) SubmitApiEvents(ctx context.Context, apiEvents []*api.EventMessage) error {
+
+	// Because (queue, userId, jobSetId) may differ between events,
+	// several sequences may be necessary.
+	sequences, err := eventutil.EventSequencesFromApiEvents(apiEvents)
+	if err != nil {
+		return err
+	}
+	if len(sequences) == 0 {
+		return nil
+	}
+
+	// Incoming gRPC requests are annotated with a unique id.
+	// Pass this id through the log by adding it to the Pulsar message properties.
+	requestId := requestid.FromContextOrMissing(ctx)
+
+	// Send each sequence async. Collect any errors via ch.
+	ch := make(chan error, len(sequences))
+	defer close(ch)
+	for _, sequence := range sequences {
+		payload, err := proto.Marshal(sequence)
+		if err != nil {
+			return errors.WithStack(err)
+		}
+
+		srv.Producer.SendAsync(
+			ctx,
+			&pulsar.ProducerMessage{
+				Payload: payload,
+				Properties: map[string]string{
+					requestid.MetadataKey:                     requestId,
+					armadaevents.PULSAR_MESSAGE_TYPE_PROPERTY: armadaevents.PULSAR_CONTROL_MESSAGE,
+				},
+				Key: sequence.JobSetName,
+			},
+			// Callback on send.
+			func(_ pulsar.MessageID, _ *pulsar.ProducerMessage, err error) {
+				ch <- err
+			},
+		)
+		if err != nil {
+			err = errors.WithStack(err)
+			return err
+		}
+	}
+
+	// Flush queued messages and wait until persisted.
+	err = srv.Producer.Flush()
+	if err != nil {
+		return err
+	}
+
+	// Collect any errors experienced by the async send and return.
+	var result *multierror.Error
+	for range sequences {
+		result = multierror.Append(result, <-ch)
+	}
+	return result.ErrorOrNil()
+}
+
 // SubmitApiEvent converts an api.EventMessage into Pulsar state transition messages and publishes those to Pulsar.
 func (srv *PulsarSubmitServer) SubmitApiEvent(ctx context.Context, apiEvent *api.EventMessage) error {
-	sequence, err := PulsarSequenceFromApiEvent(apiEvent)
+	sequence, err := eventutil.EventSequenceFromApiEvent(apiEvent)
 	if err != nil {
 		return err
 	}
@@ -601,539 +660,4 @@ func (srv *PulsarSubmitServer) SubmitApiEvent(ctx context.Context, apiEvent *api
 	}
 
 	return nil
-}
-
-// PulsarSequenceFromApiEvent converts an api.EventMessage into the corresponding Pulsar event
-// and returns an EventSequence containing this single event.
-func PulsarSequenceFromApiEvent(msg *api.EventMessage) (sequence *armadaevents.EventSequence, err error) {
-	sequence = &armadaevents.EventSequence{}
-
-	switch m := msg.Events.(type) {
-	case *api.EventMessage_Submitted:
-		// Do nothing; the Pulsar submitted message is generated by the Pulsar API endpoint.
-	case *api.EventMessage_Queued:
-		// Do nothing; there's no corresponding Pulsar message.
-	case *api.EventMessage_DuplicateFound:
-		sequence.Queue = m.DuplicateFound.Queue
-		sequence.JobSetName = m.DuplicateFound.JobSetId
-
-		newJobId, err := armadaevents.ProtoUuidFromUlidString(m.DuplicateFound.JobId)
-		if err != nil {
-			return nil, err
-		}
-		oldJobId, err := armadaevents.ProtoUuidFromUlidString(m.DuplicateFound.OriginalJobId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobDuplicateDetected{
-				JobDuplicateDetected: &armadaevents.JobDuplicateDetected{
-					NewJobId: newJobId,
-					OldJobId: oldJobId,
-				},
-			},
-		})
-	case *api.EventMessage_Leased:
-		sequence.Queue = m.Leased.Queue
-		sequence.JobSetName = m.Leased.JobSetId
-
-		// Message has no KubernetesId; use the all-zeros id.
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Leased.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunLeased{
-				JobRunLeased: &armadaevents.JobRunLeased{
-					RunId:      legacyJobRunId(),
-					JobId:      jobId,
-					ExecutorId: m.Leased.ClusterId,
-				},
-			},
-		})
-	case *api.EventMessage_LeaseReturned:
-		sequence.Queue = m.LeaseReturned.Queue
-		sequence.JobSetName = m.LeaseReturned.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.LeaseReturned.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.LeaseReturned.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunErrors{
-				JobRunErrors: &armadaevents.JobRunErrors{
-					RunId: runId,
-					JobId: jobId,
-					Errors: []*armadaevents.Error{
-						{
-							Terminal: true, // EventMessage_LeaseReturned indicates a failed job run.
-							Reason: &armadaevents.Error_PodError{
-								PodError: &armadaevents.PodError{
-									ObjectMeta: &armadaevents.ObjectMeta{
-										ExecutorId:   m.LeaseReturned.ClusterId,
-										Namespace:    "",
-										Name:         "",
-										KubernetesId: m.LeaseReturned.KubernetesId,
-										Annotations:  nil,
-										Labels:       nil,
-									},
-									PodNumber: m.LeaseReturned.PodNumber,
-									Message:   m.LeaseReturned.Reason,
-								},
-							},
-						},
-					},
-				},
-			},
-		})
-	case *api.EventMessage_LeaseExpired:
-		sequence.Queue = m.LeaseExpired.Queue
-		sequence.JobSetName = m.LeaseExpired.JobSetId
-
-		// Message has no KubernetesId; use the all-zeros id.
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.LeaseExpired.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunErrors{
-				JobRunErrors: &armadaevents.JobRunErrors{
-					RunId: legacyJobRunId(),
-					JobId: jobId,
-					Errors: []*armadaevents.Error{
-						{
-							Terminal: true, // EventMessage_LeaseExpired indicates a failed job run.
-							Reason: &armadaevents.Error_LeaseExpired{
-								LeaseExpired: &armadaevents.LeaseExpired{},
-							},
-						},
-					},
-				},
-			},
-		})
-	case *api.EventMessage_Pending:
-		sequence.Queue = m.Pending.Queue
-		sequence.JobSetName = m.Pending.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Pending.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.Pending.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunAssigned{
-				JobRunAssigned: &armadaevents.JobRunAssigned{
-					RunId: runId,
-					JobId: jobId,
-					ResourceInfos: []*armadaevents.KubernetesResourceInfo{
-						{
-							ObjectMeta: &armadaevents.ObjectMeta{
-								KubernetesId: m.Pending.KubernetesId,
-								Name:         m.Pending.PodName,
-								Namespace:    m.Pending.PodNamespace,
-								ExecutorId:   m.Pending.ClusterId,
-							},
-							Info: &armadaevents.KubernetesResourceInfo_PodInfo{
-								PodInfo: &armadaevents.PodInfo{
-									PodNumber: m.Pending.PodNumber,
-								},
-							},
-						},
-					},
-				},
-			},
-		})
-	case *api.EventMessage_Running:
-		sequence.Queue = m.Running.Queue
-		sequence.JobSetName = m.Running.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Running.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.Running.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunRunning{
-				JobRunRunning: &armadaevents.JobRunRunning{
-					RunId: runId,
-					JobId: jobId,
-					ResourceInfos: []*armadaevents.KubernetesResourceInfo{
-						{
-							ObjectMeta: &armadaevents.ObjectMeta{
-								Namespace:    m.Running.NodeName,
-								Name:         m.Running.PodName,
-								KubernetesId: m.Running.KubernetesId,
-								// TODO: These should be included.
-								Annotations: nil,
-								Labels:      nil,
-							},
-							Info: &armadaevents.KubernetesResourceInfo_PodInfo{
-								PodInfo: &armadaevents.PodInfo{
-									NodeName:  m.Running.NodeName,
-									PodNumber: m.Running.PodNumber,
-								},
-							},
-						},
-					},
-				},
-			},
-		})
-	case *api.EventMessage_UnableToSchedule:
-		sequence.Queue = m.UnableToSchedule.Queue
-		sequence.JobSetName = m.UnableToSchedule.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.UnableToSchedule.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.UnableToSchedule.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunErrors{
-				JobRunErrors: &armadaevents.JobRunErrors{
-					RunId: runId,
-					JobId: jobId,
-					Errors: []*armadaevents.Error{
-						{
-							Terminal: true, // EventMessage_UnableToSchedule indicates a failed job.
-							Reason: &armadaevents.Error_PodUnschedulable{
-								PodUnschedulable: &armadaevents.PodUnschedulable{
-									ObjectMeta: &armadaevents.ObjectMeta{
-										ExecutorId:   m.UnableToSchedule.ClusterId,
-										Namespace:    m.UnableToSchedule.PodNamespace,
-										Name:         m.UnableToSchedule.PodName,
-										KubernetesId: m.UnableToSchedule.KubernetesId,
-										Annotations:  nil,
-										Labels:       nil,
-									},
-									Message:   m.UnableToSchedule.Reason,
-									NodeName:  m.UnableToSchedule.NodeName,
-									PodNumber: m.UnableToSchedule.PodNumber,
-								},
-							},
-						},
-					},
-				},
-			},
-		})
-	case *api.EventMessage_Failed:
-		sequence.Queue = m.Failed.Queue
-		sequence.JobSetName = m.Failed.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Failed.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.Failed.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		// EventMessage_Failed contains one error for each container.
-		// Convert each of these to the corresponding Pulsar error.
-		containerErrors := make([]*armadaevents.ContainerError, 0, len(m.Failed.ContainerStatuses))
-		for _, st := range m.Failed.ContainerStatuses {
-			containerError := &armadaevents.ContainerError{
-				ExitCode: st.ExitCode,
-				Message:  st.Message,
-				Reason:   st.Reason,
-				ObjectMeta: &armadaevents.ObjectMeta{
-					ExecutorId:   m.Failed.ClusterId,
-					Namespace:    m.Failed.PodNamespace,
-					Name:         st.Name,
-					KubernetesId: "", // only the id of the pod is stored in the failed message
-				},
-			}
-
-			// Legacy messages encode the reason as an enum, whereas Pulsar uses objects.
-			switch m.Failed.Cause {
-			case api.Cause_DeadlineExceeded:
-				containerError.KubernetesReason = &armadaevents.ContainerError_DeadlineExceeded_{}
-			case api.Cause_Error:
-				containerError.KubernetesReason = &armadaevents.ContainerError_Error{}
-			case api.Cause_Evicted:
-				containerError.KubernetesReason = &armadaevents.ContainerError_Evicted_{}
-			case api.Cause_OOM:
-				containerError.KubernetesReason = &armadaevents.ContainerError_DeadlineExceeded_{}
-			default:
-				return nil, errors.WithStack(&armadaerrors.ErrInvalidArgument{
-					Name:    "Cause",
-					Value:   m.Failed.Cause,
-					Message: "Unknown cause",
-				})
-			}
-
-			containerErrors = append(containerErrors, containerError)
-		}
-
-		// Event indicating the job run failed.
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunErrors{
-				JobRunErrors: &armadaevents.JobRunErrors{
-					RunId: runId,
-					JobId: jobId,
-					Errors: []*armadaevents.Error{
-						{
-							Terminal: true,
-							Reason: &armadaevents.Error_PodError{
-								PodError: &armadaevents.PodError{
-									ObjectMeta: &armadaevents.ObjectMeta{
-										ExecutorId:   m.Failed.ClusterId,
-										Namespace:    m.Failed.PodNamespace,
-										Name:         m.Failed.PodName,
-										KubernetesId: m.Failed.KubernetesId,
-									},
-									Message:         m.Failed.Reason,
-									NodeName:        m.Failed.NodeName,
-									PodNumber:       m.Failed.PodNumber,
-									ContainerErrors: containerErrors,
-								},
-							},
-						},
-					},
-				},
-			},
-		})
-
-		// Event indicating that the job as a whole failed.
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobErrors{
-				JobErrors: &armadaevents.JobErrors{
-					JobId: jobId,
-					Errors: []*armadaevents.Error{
-						{
-							Terminal: true,
-							Reason: &armadaevents.Error_MaxRunsExceeded{
-								MaxRunsExceeded: &armadaevents.MaxRunsExceeded{},
-							},
-						},
-					},
-				},
-			},
-		})
-	case *api.EventMessage_Succeeded:
-		sequence.Queue = m.Succeeded.Queue
-		sequence.JobSetName = m.Succeeded.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Succeeded.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.Succeeded.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunSucceeded{
-				JobRunSucceeded: &armadaevents.JobRunSucceeded{
-					RunId: runId,
-					JobId: jobId,
-				},
-			},
-		})
-	case *api.EventMessage_Reprioritized:
-		sequence.Queue = m.Reprioritized.Queue
-		sequence.JobSetName = m.Reprioritized.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Reprioritized.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		priority, err := eventutil.LogSubmitPriorityFromApiPriority(m.Reprioritized.NewPriority)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_ReprioritisedJob{
-				ReprioritisedJob: &armadaevents.ReprioritisedJob{
-					JobId:    jobId,
-					Priority: priority,
-				},
-			},
-		})
-	case *api.EventMessage_Cancelling:
-		// Do nothing; there's no corresponding Pulsar message.
-	case *api.EventMessage_Cancelled:
-		sequence.Queue = m.Cancelled.Queue
-		sequence.JobSetName = m.Cancelled.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Cancelled.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_CancelledJob{
-				CancelledJob: &armadaevents.CancelledJob{
-					JobId: jobId,
-				},
-			},
-		})
-	case *api.EventMessage_Terminated:
-		// EventMessage_Terminated is generated when lease renewal fails. One such event is generated per pod.
-		// Hence, we translate these to PodError with an empty ContainerErrors.
-
-		sequence.Queue = m.Terminated.Queue
-		sequence.JobSetName = m.Terminated.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Terminated.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.Terminated.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_JobRunErrors{
-				JobRunErrors: &armadaevents.JobRunErrors{
-					RunId: runId,
-					JobId: jobId,
-					Errors: []*armadaevents.Error{
-						{
-							Terminal: true,
-							Reason: &armadaevents.Error_PodError{
-								PodError: &armadaevents.PodError{
-									ObjectMeta: &armadaevents.ObjectMeta{
-										ExecutorId:   m.Terminated.ClusterId,
-										Namespace:    m.Terminated.PodNamespace,
-										Name:         m.Terminated.PodName,
-										KubernetesId: m.Terminated.KubernetesId,
-									},
-									PodNumber: m.Terminated.PodNumber,
-									Message:   m.Terminated.Reason,
-								},
-							},
-						},
-					},
-				},
-			},
-		})
-	case *api.EventMessage_Utilisation:
-		sequence.Queue = m.Utilisation.Queue
-		sequence.JobSetName = m.Utilisation.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.Utilisation.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.Utilisation.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Created: &m.Utilisation.Created,
-			Event: &armadaevents.EventSequence_Event_ResourceUtilisation{
-				ResourceUtilisation: &armadaevents.ResourceUtilisation{
-					RunId: runId,
-					JobId: jobId,
-					ResourceInfo: &armadaevents.KubernetesResourceInfo{
-						ObjectMeta: &armadaevents.ObjectMeta{
-							ExecutorId:   m.Utilisation.ClusterId,
-							KubernetesId: m.Utilisation.KubernetesId,
-							Namespace:    m.Utilisation.PodNamespace,
-							Name:         m.Utilisation.PodName,
-						},
-						Info: &armadaevents.KubernetesResourceInfo_PodInfo{
-							PodInfo: &armadaevents.PodInfo{
-								NodeName:  m.Utilisation.NodeName,
-								PodNumber: m.Utilisation.PodNumber,
-							},
-						},
-					},
-					MaxResourcesForPeriod: m.Utilisation.MaxResourcesForPeriod,
-					TotalCumulativeUsage:  m.Utilisation.TotalCumulativeUsage,
-				},
-			},
-		})
-	case *api.EventMessage_IngressInfo:
-		// Later, ingress info should be bundled with the JobRunRunning message.
-		// For now, we create a special message that exists only for compatibility with the legacy messages.
-
-		sequence.Queue = m.IngressInfo.Queue
-		sequence.JobSetName = m.IngressInfo.JobSetId
-
-		jobId, err := armadaevents.ProtoUuidFromUlidString(m.IngressInfo.JobId)
-		if err != nil {
-			return nil, err
-		}
-
-		runId, err := armadaevents.ProtoUuidFromUuidString(m.IngressInfo.KubernetesId)
-		if err != nil {
-			return nil, err
-		}
-
-		sequence.Events = append(sequence.Events, &armadaevents.EventSequence_Event{
-			Event: &armadaevents.EventSequence_Event_StandaloneIngressInfo{
-				StandaloneIngressInfo: &armadaevents.StandaloneIngressInfo{
-					RunId: runId,
-					JobId: jobId,
-					ObjectMeta: &armadaevents.ObjectMeta{
-						ExecutorId:   m.IngressInfo.ClusterId,
-						Namespace:    m.IngressInfo.PodNamespace, // We assume the ingress was created with the same namespace as the pod
-						KubernetesId: m.IngressInfo.KubernetesId,
-					},
-					IngressAddresses: m.IngressInfo.IngressAddresses,
-					NodeName:         m.IngressInfo.NodeName,
-					PodName:          m.IngressInfo.PodName,
-					PodNumber:        m.IngressInfo.PodNumber,
-					PodNamespace:     m.IngressInfo.PodNamespace,
-				},
-			},
-		})
-	case *api.EventMessage_Reprioritizing:
-		// Do nothing; there's no corresponding Pulsar message.
-	case *api.EventMessage_Updated:
-		// Do nothing; we're not allowing arbitrary job updates.
-	default:
-		err = &armadaerrors.ErrInvalidArgument{
-			Name:    "msg",
-			Value:   msg,
-			Message: "received unsupported api message",
-		}
-		err = errors.WithStack(err)
-		return nil, err
-	}
-
-	return sequence, nil
-}
-
-func legacyJobRunId() *armadaevents.Uuid {
-	jobRunId, err := armadaevents.ProtoUuidFromUlidString(LEGACY_RUN_ID)
-	if err != nil {
-		panic(err)
-	}
-	return jobRunId
 }
